requirements.txt
gymnasium>=0.29.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
torch>=2.0.0
streamlit>=1.30.0
imageio>=2.31.0
pygame>=2.6.0
